leaderBoard/backend/requirements.txt
asgiref==3.3.2
astroid==2.4.2
certifi==2020.6.20
Django==3.2.6
djangorestframework==3.11.0
flake8==3.8.3
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pylint-django==2.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0